Migrated javadoc-style annotation to Maven Plugin Annotations.
The two main reasons for this move:

1. It is partially checked by the compiler

2. It makes interoperability easier, as the Java annotations were preserved in the classfiles whereas the javadoc-annoations were not.
E.g. The bloop project has a maven plugin which inherits from some base classes of this plugin to read it's configuration.
After this change, the maven plugin build just works, but before, it couldn't see any @parameter and @component comments, which was a real pain.