Un-break the doc-jar goal.
Beginning with d9f45d0, the doc-jar goal (ScalaDocJarMojo) was not passing a classpath to scaladoc. When running on Java 8, scaladoc reported errors wherever any symbol from a dependency is mentioned (e.g. “object log4s is not a member of package org”, even though log4s is supposed to be on the classpath). When running on Java 10, scaladoc suffered an immediate fatal error instead (“object scala in compiler mirror not found”).

The problem was that ScalaDocJarMojo did not declare that it `requiresDependencyResolution`. In the legacy Javadoc-style annotations for Maven plugins, `@requiresDependencyResolution` is a separate, inherited annotation, so ScalaDocJarMojo inherited it from ScalaDocMojo. With the new annotations, however, `requiresDependencyResolution` is instead a field of the `@Mojo` annotation, which is *not* inherited.