Run incremental compiler in a forked process when jvmArgs are defined, close #507
Motivation:

The incremental mode which is now the default one doesn't handle jvmArgs as it can't run in a forked process.
A user might need this for example in order to increase the Xss to compile complex code with lots of method chaining and type inference.

Modification:

* introduce a new Fork class (only used for the incremental compiler for now.
* run the incremental compiler as soon as jvmArgs are defined

Result:

The default incremental compiler now properly handles jvmArgs.