fix: Handle race conditions when getting the compiler bridge
It was identified in https://github.com/davidB/scala-maven-plugin/issues/738 that a race condition could occur when building a project using multiple threads.

To mitigate the issue, I wrapped the installation of the compiler with a synchronized block on the Static instance.

As a result, if multiple threads try to install the compiler at the same time, only one thread will be able to do so. When they will get unlocked, it will check if the file was created meanwhile and if so, to will early return.