fix(deps): upgrade minimal version of maven from 3.3.9 to 3.8.1
Fix a security alert

Apache Maven will follow repositories that are defined in a dependency’s
Project Object Model (pom) which may be surprising to some users,
resulting in potential risk if a malicious actor takes over that
repository or is able to insert themselves into a position to pretend to
be that repository. Maven is changing the default behavior in 3.8.1+ to
no longer follow http (non-SSL) repository references by default. More
details available in the referenced urls. If you are currently using a
repository manager to govern the repositories used by your builds, you
are unaffected by the risks present in the legacy behavior, and are
unaffected by this vulnerability and change to default behavior. See
this link for more information about repository management:
https://maven.apache.org/repository-management.html